+++ hera-backend/src/main/java/com/hera/backend/repository/FamiliaOlfativaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface FamiliaOlfativaRepository extends JpaRepository<FamiliaOlfativa, Integer> {
+public interface FamiliaOlfativaRepository extends JpaRepository<FamiliaOlfativa, Long> {
 
     Optional<FamiliaOlfativa> findByNombre(String nombre);
 }
